refactor: move rename useRefState.ts to useStateRef.ts
